+++ bbs/bbsovl1.cpp
@@ -169,7 +169,7 @@ void send_email() {
   // TODO(rushfan): This is where we'd check for (NNNN) 
   // and add in the @NNN for the FTN networks.
 
-  int system_number, user_number;
+  uint16_t system_number, user_number;
   parse_email_info(username, &user_number, &system_number);
   grab_quotes(nullptr, nullptr);
   if (user_number || system_number) {
